+++ src/main/java/net/geforcemods/securitycraft/blocks/DisplayCaseBlock.java
@@ -160,7 +160,7 @@ public class DisplayCaseBlock extends OwnableBlock implements SimpleWaterloggedB
 					activate(be);
 				}
 				else if (!PlayerUtils.isHoldingItem(player, SCContent.CODEBREAKER, hand))
-					be.openPasswordGUI(player);
+					be.openPasswordGUI(level, pos, be.getOwner(), player);
 			}
 		}