+++ src/main/java/net/geforcemods/securitycraft/blocks/KeyPanelBlock.java
@@ -112,7 +112,7 @@ public class KeyPanelBlock extends OwnableBlock implements SimpleWaterloggedBloc
 				activate(state, level, pos, be.getSignalLength());
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.CODEBREAKER, hand))
-				be.openPasswordGUI(player);
+				be.openPasswordGUI(level, pos, be.getOwner(), player);
 		}
 
 		return InteractionResult.SUCCESS;