+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadChestBlock.java
@@ -106,7 +106,7 @@ public class KeypadChestBlock extends ChestBlock {
 				activate(state, level, pos, player);
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.CODEBREAKER, hand))
-				be.openPasswordGUI(player);
+				be.openPasswordGUI(level, pos, be.getOwner(), player);
 		}
 
 		return InteractionResult.SUCCESS;