+++ Changelog.md
@@ -14,6 +14,7 @@
 - API: Changed LinkedAction to an interface and the enum values to records for easier usability
 - API: Several methods in LinkableBlockEntity were changed to account for the LinkedAction change
 - API: LinkedAction has been renamed to ILinkedAction
+- API: Changes to ICodebreakable and IPasswordProtected to cut down on duplicate code and make them easier and more clear to use
 - Fix: Nether portals can replace reinforced blocks when generating
 - Fix: Crash when trying to disguise a block as/project some blocks that don't always have a block entity associated with them
 - Fix: Can't place blocks on blocks locked by a Sonic Security System